refactor: NewsPage로 API관련 코드 적용
